Have got draft implementation of MVC working. In the currently implemented system the controller has a model, and the model has a view. The controller registers event handlers on its model's view.
The view contain by the model must be made transient for serialization to work, therefore after a dragboard has been used to copy data the model, view and controller need be reconstructed.

The dragboard will not provide a reference to the originating object, only a copy of its data, therefore a static object can be used in place.