Recreated all the different types of links as various link views, there is just the one model that has a link type property. There is some inconsistancy between having a Type Of link and a link Has A type; although it seems to work ok for now.
The various views have lots of similar code in them, but with many minor variations. There may be some optimization possible within the OO model. The difficulty is that different path elements are used, and the super type PathElement itself has no x and y properties that can be set.