When a connected node changes from parent to child, reverseDirection Binding is set, path elements start and end are set to listen to opposing nodes etc.
However, there is a problem with values being updated. For example, move node1 below node2 and relation direction changes, but node2 position remains the same, so change listener is not fired and path is incorrectly set.

Need to sort out properties from bindings and pay attention to eager and lazy evaluations.